pyproject.toml
[tool.poetry]
name = "wt-activity"
version = "0.1.0"
description = ""
authors = ["ikoposhilov <ikoposhilov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
SQLAlchemy = "^2.0.36"
psycopg2 = "^2.9.10"
uvicorn = "^0.32.0"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.0"
markdown2 = "^2.5.3"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
passlib = "1.7.4"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
